+++ src/channel.ts
@@ -22,6 +22,7 @@ import type {
   ChannelFilters,
   ChannelMemberAPIResponse,
   ChannelMemberResponse,
+  ChannelPushPreference,
   ChannelQueryOptions,
   ChannelResponse,
   ChannelUpdateOptions,
@@ -57,7 +58,6 @@ import type {
   PinnedMessagePaginationOptions,
   PinnedMessagesSort,
   PollVoteData,
-  PushPreference,
   QueryChannelAPIResponse,
   QueryMembersOptions,
   Reaction,
@@ -113,7 +113,7 @@ export class Channel {
   lastTypingEvent: Date | null;
   isTyping: boolean;
   disconnected: boolean;
-  push_preferences?: PushPreference;
+  push_preferences?: ChannelPushPreference;
   public readonly messageComposer: MessageComposer;
   public readonly messageReceiptsTracker: MessageReceiptsTracker;
   public readonly cooldownTimer: CooldownTimer;